backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
google-genai
pydantic==2.4.2
httpx==0.24.0